API_based_automations/travian_bot/requirements.txt
beautifulsoup4==4.13.4
Requests==2.32.3
tqdm==4.67.1
PyYAML==6.0.2
